+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## [Unreleased]
 
+## [0.2.0] - 2023-03-21
+- Add Deal list service (filtering by id);
+- Add User list service (filtering by id);
+
+
 ## [0.1.0] - 2023-03-02
 
 - Initial release